now monitor scripts don't use hardcoded codes/values
before the monitor control was using hardcoded settings which could lead
to undefined behaviour on other monitors. now should use adresses
specific to the monitor in the system running the script

input switch script still uses hardcoded values but now checks the model
of the monitor. should be safe to run.